requirements.txt
langchain
langchain-community
langchain-groq
neo4j
networkx

python-dotenv
scipy
google-generativeai
pytesseract
PyPDF2
chromadb
faiss-cpu
torch
transformers

matplotlib
gradio
streamlit